pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nlp-domain-alignment"
version = "0.1.0"
description = "Domain alignment pipeline for vocational programs and job postings"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    # Core data science
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scipy>=1.9.0",
    
    # Machine learning
    "scikit-learn>=1.1.0",
    "gensim>=4.0.0",
    
    # Data processing
    "pyarrow>=8.0.0",
    "beautifulsoup4>=4.11.0",
    "nltk>=3.7",
    "spacy>=3.4.0",
    
    # Visualization
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    
    # Embeddings and dimensionality reduction
    "umap-learn>=0.5.0",
    
    # Configuration and utilities
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
    "joblib>=1.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
target-version = "py38"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
]
ignore = [
    "E501",  # line too long, handled by black
]

